+++ meandmyteam/settings.py
@@ -37,7 +37,7 @@ INSTALLED_APPS = [
     'django.contrib.staticfiles',
     'widget_tweaks',
     'website',
-    #'surveys',
+    'surveys',
     'payments',
     'django_countries',
     'phonenumber_field',
