+++ website/forms.py
@@ -8,6 +8,7 @@ from website.models import Organization
 from django_countries.widgets import CountrySelectWidget
 from django_countries.fields import CountryField
 from phonenumber_field.formfields import PhoneNumberField
+from phonenumber_field.widgets import PhoneNumberPrefixWidget
 
 class SignUpForm(forms.Form):
     #User
@@ -148,7 +149,8 @@ class EditAccountForm(forms.Form):
 
     #Organization
     name = forms.CharField(max_length = 255, label="Organization name", widget=forms.TextInput(attrs={}))
-    phone = PhoneNumberField(max_length=30, required=False, label="Phone incl.country code (e.g. +01)", widget=forms.TextInput(attrs={'placeholder':'Example: +12125552368'}))
+    phone = PhoneNumberField(max_length=30, required=False, label="Phone incl.country code (e.g. +01)", widget=forms.TextInput(attrs={'placeholder':'Example: +12125552368'})) # widget=PhoneNumberPrefixWidget(attrs={})
+
     address_line_1 = forms.CharField(max_length = 255, label="Street address", widget=forms.TextInput(attrs={}))
     address_line_2 =forms.CharField(max_length = 255, label="", required=False, widget=forms.TextInput(attrs={}))
     zip_code = forms.CharField(max_length = 20, label="Zip", widget=forms.TextInput(attrs={}))
